Scrapper/requirements.txt
Augmentor
chromedriver-binary-auto
beautifulsoup4
pandas
requests
selenium
tensorflow==1.15.0
nudenet
html5lib
aiohttp